+++ src/index.js
@@ -1,10 +1,3 @@
-/**
- * function-batch
- *
- * Copyright © 2016 Michael Dougall. All rights reserved.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.txt file in the root directory of this source tree.
- */
+// @flow
 
-export default from './functionBatch.js';
+export default './functionBatch.js';