+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased][]
+### Added
+- Function to return a promise that is resolved when the original function returns a value
+
 ## [1.0.0][] - 2017-04-20
 ### Added
 - Flowtypes to npm module